+++ Options/WindturbineOptions.cs
@@ -5,7 +5,7 @@ public class WindturbineOptions
     public string Name { get; set; } = null!;
 
     public int Upkeep { get; set; }
-    public int Maximumwind { get; set; }
+    public float Maximumwind { get; set; }
     public int Production { get; set; }
     public int Noisepollution { get; set; }
